+++ Application/FileConverter/ConversionJobs/ConversionJob_Ico.cs
@@ -11,8 +11,17 @@ namespace FileConverter.ConversionJobs
         private ConversionJob pngConversionJob;
         private ConversionJob icoConversionJob;
 
-        public ConversionJob_Ico(ConversionPreset conversionPreset) : base(conversionPreset)
+        public ConversionJob_Ico(ConversionPreset conversionPreset, string inputFilePath) : base(conversionPreset, inputFilePath)
         {
+            this.IsCancelable = true;
+        }
+
+        public override void Cancel()
+        {
+            base.Cancel();
+
+            this.pngConversionJob.Cancel();
+            this.icoConversionJob.Cancel();
         }
 
         protected override void Initialize()
@@ -34,11 +43,11 @@ namespace FileConverter.ConversionJobs
             intermediatePreset.SetSettingsValue(ConversionPreset.ConversionSettingKeys.ImageClampSizePowerOf2, "True");
             intermediatePreset.SetSettingsValue(ConversionPreset.ConversionSettingKeys.ImageMaximumSize, "256");
             this.pngConversionJob = ConversionJobFactory.Create(intermediatePreset, this.InputFilePath);
-            this.pngConversionJob.PrepareConversion(this.InputFilePath, this.intermediateFilePath);
+            this.pngConversionJob.PrepareConversion(this.intermediateFilePath);
 
             // Convert png file into ico.
-            this.icoConversionJob = new ConversionJob_FFMPEG(this.ConversionPreset);
-            this.icoConversionJob.PrepareConversion(this.intermediateFilePath, this.OutputFilePath);
+            this.icoConversionJob = new ConversionJob_FFMPEG(this.ConversionPreset, this.intermediateFilePath);
+            this.icoConversionJob.PrepareConversion(this.OutputFilePath);
         }
 
         protected override void Convert()