+++ Application/FileConverter/Windows/ConversionSettingsTemplateSelector.cs
@@ -91,6 +91,12 @@ namespace FileConverter.Windows
             set;
         }
 
+        public DataTemplate WebpSettingsDataTemplate
+        {
+            get;
+            set;
+        }
+
         public override DataTemplate SelectTemplate(object item, DependencyObject container)
         {
             if (item == null)
@@ -140,6 +146,9 @@ namespace FileConverter.Windows
 
                 case OutputType.Webm:
                     return this.WebmSettingsDataTemplate;
+
+                case OutputType.Webp:
+                    return this.WebpSettingsDataTemplate;
             }
 
             return this.DefaultDataTemplate;